+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/messenger.php
@@ -9,5 +9,10 @@ $container->loadFromExtension('framework', [
             FooMessage::class => ['sender.bar', 'sender.biz'],
             BarMessage::class => 'sender.foo',
         ],
+        'transports' => [
+            'sender.biz' => 'null://',
+            'sender.bar' => 'null://',
+            'sender.foo' => 'null://',
+        ],
     ],
 ]);