+++ src/Symfony/Component/Console/CHANGELOG.md
@@ -16,6 +16,7 @@ CHANGELOG
 4.4.0
 -----
 
+ * deprecated finding hidden commands using an abbreviation, use the full name instead
  * added `Question::setTrimmable` default to true to allow the answer to be trimmed
  * added method `preventRedrawFasterThan()` and `forceRedrawSlowerThan()` on `ProgressBar`
  * `Application` implements `ResetInterface`